updated with lauren's comments
