+++ src/services/measure.service.js
@@ -17,7 +17,7 @@ const {
   assembleCollectionBundleFromMeasure,
   getQueryFromReference
 } = require('../util/bundleUtils');
-const { getPatientDataCollectionBundle } = require('../util/patientUtils');
+const { getPatientDataCollectionBundle, retrievePatientIds } = require('../util/patientUtils');
 const {
   addPendingBulkImportRequest,
   findOneResourceWithQuery,
@@ -335,7 +335,7 @@ const careGaps = async (args, { req }) => {
       measurementPeriodStart: periodStart,
       measurementPeriodEnd: periodEnd
     });
-    const patientIds = await retrievePatients(query);
+    const patientIds = await retrievePatientIds(query);
 
     let patientBundles = patientIds.map(async m => {
       return getPatientDataCollectionBundle(`Patient/${m}`, dataReq.results.dataRequirement);
@@ -375,26 +375,6 @@ const careGaps = async (args, { req }) => {
   };
   logger.info('Successfully generated $care-gaps report');
   return responseParameters;
-};
-
-const retrievePatients = async ({ subject, organization }) => {
-  let referencedObject;
-  const reference = (subject || organization).split('/');
-  if (reference[0] !== 'Patient') {
-    referencedObject = await findResourceById(reference[1], reference[0]);
-    if (!referencedObject) {
-      throw new ResourceNotFoundError(`No resource found in collection: ${reference[0]}, with id: ${reference[1]}.`);
-    }
-  }
-
-  if (reference[0] === 'Group') {
-    return referencedObject.member.map(m => m.entity.reference);
-  } else if (reference[0] === 'Patient') {
-    return [subject];
-  } else {
-    const patients = await findResourcesWithQuery({ 'managingOrganization.identifier.value': organization }, 'Patient');
-    return patients.map(e => e.id);
-  }
 };
 
 /**
