+++ src/util/patientUtils.js
@@ -71,4 +71,31 @@ async function getPatientData(patientId, dataRequirements) {
   return data;
 }
 
-module.exports = { getPatientDataCollectionBundle, getPatientDataSearchSetBundle, getPatientData };
+/**
+ * Takes in a care-gaps query and retrieves the ids of the patients to be run against the measure
+ * NOTE: Currently, this function assumes ResourceType/ResourceId-style references
+ * @param {string} subject A reference to either the patient or group resource to run against a measure
+ * @param {string} organization A reference to an organization that
+ * @returns {Array} an array of patient ids
+ */
+const retrievePatientIds = async ({ subject, organization }) => {
+  let referencedObject;
+  const reference = (subject || organization).split('/');
+  if (reference[0] !== 'Patient') {
+    referencedObject = await findResourceById(reference[1], reference[0]);
+    if (!referencedObject) {
+      throw new ResourceNotFoundError(`No resource found in collection: ${reference[0]}, with id: ${reference[1]}.`);
+    }
+  }
+
+  if (reference[0] === 'Group') {
+    return referencedObject.member.map(m => m.entity.reference.split('/')[1]);
+  } else if (reference[0] === 'Patient') {
+    return [subject.split('/')[1]];
+  } else {
+    const patients = await findResourcesWithQuery({ 'managingOrganization.identifier.value': organization }, 'Patient');
+    return patients.map(e => e.id);
+  }
+};
+
+module.exports = { getPatientDataCollectionBundle, getPatientDataSearchSetBundle, getPatientData, retrievePatientIds };
